+++ src/main/java/com/cbnuccc/cbnuccc/Repository/UserJpaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cbnuccc.cbnuccc.Model.MyUser;
 
-public interface UserJpaRepository extends JpaRepository<MyUser, Integer> {
+public interface UserJpaRepository extends JpaRepository<MyUser, Long> {
     Optional<MyUser> findByEmail(String email);
 
     Optional<MyUser> findByUuid(UUID uuid);
